// workspaces/topology/plugins/topology/src/types/pods.ts
/*
 * Copyright 2024 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { V1Pod } from '@kubernetes/client-node';

import { K8sWorkloadResource } from './types';

export type OverviewItemAlerts = {
  [key: string]: {
    message: string;
    severity: string;
  };
};

export type PodControllerOverviewItem = {
  alerts?: OverviewItemAlerts;
  revision?: number;
  obj?: K8sWorkloadResource;
  phase?: string;
  pods?: V1Pod[];
};

export type PodRCData = {
  current?: PodControllerOverviewItem;
  previous?: PodControllerOverviewItem;
  obj: K8sWorkloadResource;
  isRollingOut?: boolean;
  pods: V1Pod[];
};